// src/types/database.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5";
  };
  public: {
    Tables: {
// ...
      hotel_recommended_places: {
        Row: {
          address: string;
          created_at: string | null;
          created_by: string | null;
          description: string | null;
          hotel_id: string;
          id: string;
          is_active: boolean;
          latitud: number | null;
          longitud: number | null;
          place_name: string;
          updated_at: string | null;
        };
        Insert: {
          address: string;
          created_at?: string | null;
          created_by?: string | null;
          description?: string | null;
          hotel_id: string;
          id?: string;
          is_active?: boolean;
          latitud?: number | null;
          longitud?: number | null;
          place_name: string;
          updated_at?: string | null;
        };
        Update: {
          address?: string;
          created_at?: string | null;
          created_by?: string | null;
          description?: string | null;
          hotel_id?: string;
          id?: string;
          is_active?: boolean;
          latitud?: number | null;
          longitud?: number | null;
          place_name?: string;
          updated_at?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "hotel_recommended_places_created_by_fkey";
            columns: ["created_by"];
            isOneToOne: false;
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "hotel_recommended_places_hotel_id_fkey";
            columns: ["hotel_id"];
            isOneToOne: false;
            referencedRelation: "hotels";
            referencedColumns: ["id"];
          }
        ];
      };
// ...
    };
// ...
  };
};